+++ client/src/components/admin/Dashboard.tsx
@@ -1,12 +1,28 @@
-import { FC } from "react";
+import { useState, useEffect, useCallback } from "react";
 import AddLand from "./AddLand";
 import Lands from "./Lands";
+import { RouteComponentProps } from "@reach/router";
+import axios from "axios";
 
-const Dashboard: FC = () => {
+const Dashboard = (props: RouteComponentProps) => {
+  const [lands, setLands] = useState([]);
+  const getLands = useCallback(async () => {
+    try {
+      const { data } = await axios.get(
+        `${process.env.REACT_APP_SERVER_ADDRESS}/admin/lands`
+      );
+      setLands(data.lands);
+    } catch (error) {
+      console.log(error.message);
+    }
+  }, []);
+  useEffect(() => {
+    getLands();
+  }, [getLands]);
   return (
     <div>
-      <AddLand />
-      <Lands />
+      <AddLand getLands={() => getLands()} />
+      <Lands lands={lands} getLands={() => getLands()} />
     </div>
   );
 };