+++ server/src/routes/admin.router.ts
@@ -1,10 +1,10 @@
 import express from "express";
-
-import { getLands } from "../controllers/public.controller";
+import { addLand, getLands } from "../controllers/admin.controller";
 
 const router = express.Router();
 
 router.get("/lands", getLands);
+router.post("/land/add", addLand);
 // router.post("/land/:landid", loginAdmin);
 
 export default router;