+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Backend
 
-backend
-
-none
+- routes for adding and fetching lands added
+- models, and interfaces of lands, assets and admin added
+- controllers added for land
+- added public route
 
 # Frontend
+
+- added admin page and form to add lands
+- added land lists in admin page
+- added api integration in homepage for lands
 
 added image gallery view in frontend property card